+++ src/components/button/UploadFileBtn.jsx
@@ -3,15 +3,16 @@ import "./UploadFileBtn.css";
 export default function UploadFileBtn(props) {
     return (
         <>
-            <label 
-            htmlFor={props.forAndId} 
-            className={`UploadFileLabel ${props.type}`}
-            style={{right: `${props.right}`, 
+            <label
+                htmlFor={props.forAndId}
+                className={`UploadFileLabel ${props.type}`}
+                style={{
+                    right: `${props.right}`,
                     bottom: `${props.bottom}`,
-                    position: `${props.position}`
-            }}/>
+                    position: `${props.position}`,
+                }}
+            />
             <input type="file" id={props.forAndId} />
         </>
     );
 }
-
