+++ src/components/common/BasicProfileImg.jsx
@@ -1,5 +1,5 @@
 import "./BasicProfileImg.css";
-import BasicProfile from "../assets/basic-profile-img.svg";
+import BasicProfile from "../../assets/basic-profile-img.svg";
 
 export default function BasicProfileImg(props) {
     return (
