+++ src/components/common/TopBasicNav.css
@@ -10,11 +10,11 @@
 .prevBtn {
     width: 22px;
     height: 22px;
-    background: url("../assets/icon/icon-arrow-left.svg");
+    background: url("../../assets/icon/icon-arrow-left.svg");
 }
 
 .moreBtn {
     width: 22px;
     height: 22px;
-    background: url("../assets/icon/icon-\ more-vertical.svg");
+    background: url("../../assets/icon/icon-\ more-vertical.svg");
 }
