+++ src/pages/FollowingList.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import FollowingContainer from "../../components/follow/FollowingContainer";
 import FollowingHeader from "../../components/follow/FollowingHeader";
-import "./FollowerList.css";
+import "../pages/style/FollowerList.css";
 function FollowingList() {
     return (
         <div className="followerWrap">
