+++ src/pages/Intro.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import './Intro.css'
+import "../pages/style/Intro.css";
 
 function Intro() {
     return (
@@ -8,12 +8,22 @@ function Intro() {
             <section className="loginSection">
                 <div className="pea">완두콩</div>
                 <article className="centerWebLogin">
-                    <button className="kakao" type="button">카카오톡 계정으로 로그인</button>
-                    <button className="google" type="button">구글 계정으로 로그인</button>
-                    <button className="facebook" type="button">페이스북 계정으로 로그인</button>
+                    <button className="kakao" type="button">
+                        카카오톡 계정으로 로그인
+                    </button>
+                    <button className="google" type="button">
+                        구글 계정으로 로그인
+                    </button>
+                    <button className="facebook" type="button">
+                        페이스북 계정으로 로그인
+                    </button>
                     <p className="bottomLogin">
-                        <a id="email" href="/Login">이메일로 로그인</a>
-                        <a id="signup" href="/join">회원가입</a>
+                        <a id="email" href="/Login">
+                            이메일로 로그인
+                        </a>
+                        <a id="signup" href="/join">
+                            회원가입
+                        </a>
                     </p>
                 </article>
             </section>
