+++ src/pages/MyProfile.jsx
@@ -1,7 +1,7 @@
-import TopBasicNav from "../../components/TopBasicNav";
-import "./MyProfile.css";
-import ProfileInfo from "../../components/profile/ProfileInfo";
-import ProductContainer from "../../components/product/ProductContainer";
+import TopBasicNav from "../components/common/TopBasicNav";
+import ProductContainer from "../components/product/ProductContainer";
+import ProfileInfo from "../components/profile/ProfileInfo";
+import "../pages/style/MyProfile.css";
 
 export default function MyProfile() {
     return (
