+++ src/pages/style/Intro.css
@@ -1,65 +1,67 @@
-@import url('https://fonts.googleapis.com/css2?family=Gamja+Flower&display=swap');
-        .pea{
-            background-image: url(../assets/pea-nocolor.png);
-            background-repeat: no-repeat;
-            background-size: contain;
-            text-indent: -9999px;
-            width: 200px;
-            height: 200px;
-            margin: 150px auto;
-        }
-        .loginMain{
-            background-color: var(--main-color);
-            position: relative;
-            height: 100vh;
-        }
-        .centerWebLogin{
-            position: absolute;
-            bottom: 0;
-            width: 100%;
-            -webkit-box-sizing: border-box;
-                    box-sizing: border-box;
-            background-color: #ffffff;
-            padding: 50px 34px;
-            margin-top: 70px;
-            border-top-left-radius: 20px;
-            border-top-right-radius: 20px;
-        }
-        .bottomLogin a{
-            text-decoration: none;
-            font-size: 12px;
-        }
-        button{
-            width: 100%;
-        }
-        .kakao, .google, .facebook{
-            color: var(--sub-font-color);
-            margin-bottom: 10px;
-            padding: 13px 0;
-            border-radius: 45px;
-            font-size: 14px;
-            background-repeat: no-repeat;
-            background-position: 15px center;
-        }
-        .kakao{
-            border: 1px solid #F2C94C;
-            background-image: url(../assets/message-circle.png);
-        }
-        .google{
-            border: 1px solid #767676;
-            background-image: url(../assets/google.png);
-        }
-        .facebook{
-            border: 1px solid #2D9CDB;
-            background-image: url(../assets/facebook.png);
-        }
-        .bottomLogin{
-            text-align: center;
-            margin: 20px auto;
-        }
-        #email{
-            margin-right: 20px;
-        }
-        #singup{
-            margin-left: 20px;
-        }
+@import url("https://fonts.googleapis.com/css2?family=Gamja+Flower&display=swap");
+.pea {
+    background-image: url(../../assets/pea-nocolor.png);
+    background-repeat: no-repeat;
+    background-size: contain;
+    text-indent: -9999px;
+    width: 200px;
+    height: 200px;
+    margin: 150px auto;
+}
+.loginMain {
+    background-color: var(--main-color);
+    position: relative;
+    height: 100vh;
+}
+.centerWebLogin {
+    position: absolute;
+    bottom: 0;
+    width: 100%;
+    -webkit-box-sizing: border-box;
+    box-sizing: border-box;
+    background-color: #ffffff;
+    padding: 50px 34px;
+    margin-top: 70px;
+    border-top-left-radius: 20px;
+    border-top-right-radius: 20px;
+}
+.bottomLogin a {
+    text-decoration: none;
+    font-size: 12px;
+}
+button {
+    width: 100%;
+}
+.kakao,
+.google,
+.facebook {
+    color: var(--sub-font-color);
+    margin-bottom: 10px;
+    padding: 13px 0;
+    border-radius: 45px;
+    font-size: 14px;
+    background-repeat: no-repeat;
+    background-position: 15px center;
+}
+.kakao {
+    border: 1px solid #f2c94c;
+    background-image: url(../../assets/message-circle.png);
+}
+.google {
+    border: 1px solid #767676;
+    background-image: url(../../assets/google.png);
+}
+.facebook {
+    border: 1px solid #2d9cdb;
+    background-image: url(../../assets/facebook.png);
+}
+.bottomLogin {
+    text-align: center;
+    margin: 20px auto;
+}
+#email {
+    margin-right: 20px;
+}
+#singup {
+    margin-left: 20px;
+}
