+++ src/pages/style/Splash.css
@@ -1,19 +1,19 @@
-@import url('https://fonts.googleapis.com/css2?family=Gamja+Flower&display=swap');
-        .headTitle{
-            text-indent: -9999px;
-        }
-        .colorPea{
-            background-image: url(../assets/pea-color.png);
-            background-repeat: no-repeat;
-            background-size: contain;
-            text-indent: -9999px;
-            width: 200px;
-            height: 200px;
-            margin: 30% auto;
-        }
-        .mainMarket{
-            text-align: center;
-            color: #83BD75;
-            font-family: 'Gamja Flower', cursive;
-            font-size: 48px;
-        }
+@import url("https://fonts.googleapis.com/css2?family=Gamja+Flower&display=swap");
+.headTitle {
+    text-indent: -9999px;
+}
+.colorPea {
+    background-image: url(../../assets/pea-color.png);
+    background-repeat: no-repeat;
+    background-size: contain;
+    text-indent: -9999px;
+    width: 200px;
+    height: 200px;
+    margin: 30% auto;
+}
+.mainMarket {
+    text-align: center;
+    color: #83bd75;
+    font-family: "Gamja Flower", cursive;
+    font-size: 48px;
+}
